orders/requirements-backup.txt
django
django-filter
djangorestframework
django-rest-passwordreset
django-extensions
django-allauth==0.52.0
psycopg2-binary
gunicorn
pyyaml
pyjwt
requests
django-filter
importlib-metadata
redis
pytest
pytest-django
pytest-cov
model_bakery
drf-spectacular
celery
